airflow/requirements.txt
apache-airflow-providers-dbt-cloud
apache-airflow-providers-google==10.12.0
dbt-bigquery==1.10.2
google-cloud-bigquery>=3.0.0
dbt-core==1.10.13
astronomer-cosmos>=1.0.0